made results activity main for the monent for dummy data
