+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,12 @@
         </activity>
         <activity
             android:name=".activities.SearchResultsActivity"
-            android:label="@string/title_activity_search_results" />
+            android:label="@string/title_activity_search_results">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".activities.DetailActivity"
             android:label="@string/title_activity_detail" />
@@ -25,10 +30,6 @@
             android:name=".activities.SearchPagerActivity"
             android:theme="@style/SearchActivityTheme"
             android:windowSoftInputMode="adjustResize">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <meta-data
             android:name="io.fabric.ApiKey"
