+++ app/src/main/java/edu/grinnell/appdev/grinnelldirectory/activities/SearchResultsActivity.java
@@ -6,11 +6,16 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.MenuItem;
 
+import com.google.gson.Gson;
+
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.List;
 
 import edu.grinnell.appdev.grinnelldirectory.R;
 import edu.grinnell.appdev.grinnelldirectory.adapters.SearchResultsAdapter;
 import edu.grinnell.appdev.grinnelldirectory.models.Person;
+import edu.grinnell.appdev.grinnelldirectory.models.Persons;
 import edu.grinnell.appdev.grinnelldirectory.models.SimpleResult;
 
 public class SearchResultsActivity extends AppCompatActivity {
@@ -21,6 +26,8 @@ public class SearchResultsActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_searchresults);
 
+        loadPersons();
+
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         RecyclerView rvSearchResults = (RecyclerView) findViewById(R.id.rvSearchResults);
@@ -29,7 +36,10 @@ public class SearchResultsActivity extends AppCompatActivity {
         rvSearchResults.setLayoutManager(new LinearLayoutManager(this));
 
         SimpleResult result = getIntent().getParcelableExtra(SimpleResult.SIMPLE_KEY);
-        adapter.updateData(result.getPeople());
+        // REMOVED
+        //adapter.updateData(result.getPeople());
+        // FOR
+        adapter.updateData(mPeopleList);
     }
 
     @Override
@@ -42,5 +52,32 @@ public class SearchResultsActivity extends AppCompatActivity {
                 break;
         }
         return true;
+    }
+
+
+    private void loadPersons() {
+        // get multiple persons
+        String json = loadJSONFromAsset("dummyData.json");
+        Persons persons = new Gson().fromJson(json, Persons.class);
+        System.out.println(persons.getPersons().get(0).getFirstName());
+
+        // return our persons object
+        this.mPeopleList = persons.getPersons();
+    }
+
+    public String loadJSONFromAsset(String fileName) {
+        String json = null;
+        try {
+            InputStream is = this.getAssets().open(fileName);
+            int size = is.available();
+            byte[] buffer = new byte[size];
+            is.read(buffer);
+            is.close();
+            json = new String(buffer, "UTF-8");
+        } catch (IOException ex) {
+            ex.printStackTrace();
+            return null;
+        }
+        return json;
     }
 }
